Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/FileAPI/reading-data-section/filereader_readAsBinaryString-expected.txt


FAIL FileAPI Test: filereader_readAsBinaryString assert_equals: expected 1 but got 2
